Store Axel usage of this ETL script
- return useful error and suggest fix when google cloud project is not specified
  - before it would just print out error and the cleanup would remove all traces of where the command failed
- print out more echo statements so we know what's going on
- use python3 instead of python

+++ bigquery_to_gcs.sh
@@ -25,12 +25,19 @@ output_uri="gs://${output_bucket}/${output_folder_raw}/*.gz"
 echo "Extracting ${input_table} to ${output_uri} ..."
 bq extract --noprint_header --destination_format=CSV --compression=GZIP ${input_table} ${output_uri}
 
+# TODO: Update it to handle python and python3 - aliases don't work in bash scripts, so you should detect what python is available
 # BigQuery exports partitioned tables to many small files. We combine them to bigger files to import faster.
-python ethereumetl_postgres/gcs_compose.py -b ${output_bucket} -i "${output_folder_raw}/" -o "${output_folder_composed}/"
+echo "Combine BigQuery exports to bigger files so we that we can import faster."
+
+output=$(python3 ethereumetl_postgres/gcs_compose.py -b ${output_bucket} -i "${output_folder_raw}/" -o "${output_folder_composed}/" 2>&1 | tee /dev/tty)
+if [[ $output =~ "OSError: Project was not passed and could not be determined from the environment." ]]; then
+    echo -e "-e \"\033[1;37;41mERROR:\033[0m Please set your project using the following command: gcloud config set project _project_name_"
+    exit 1
+fi
 # gcloud compose has limit of 32 files in one compose operation. Try composing once more to work around this limit.
-python ethereumetl_postgres/gcs_compose.py -b ${output_bucket} -i "${output_folder_composed}/" -o "${output_folder}/"
+python3 ethereumetl_postgres/gcs_compose.py -b ${output_bucket} -i "${output_folder_composed}/" -o "${output_folder}/"
 
 # Cleanup
-
+echo "Cleaning up the raw and composed folders from the Google Cloud Bucket"
 gsutil -m rm -r "gs://${output_bucket}/${output_folder_raw}"
 gsutil -m rm -r "gs://${output_bucket}/${output_folder_composed}"
